Fix issues with unit tests
